Rotation using Slerp (lerp ftw)
I made aimRotateRate =1 and lookRotateRate =10, but that's my opinion on
what looks natural. Tweak that value around to see if you agree or
disagree on what that value should be.

Also, I changed turret to = Turret Mount instead of turret equaling the
muzzel, which ended up making the muzzel falling off the tank when I
used the new aim method.